+++ R/kaplan.R
@@ -51,6 +51,7 @@ kaplanUI <- function(id) {
     checkboxInput(ns("cumhaz"), "Show cumulative incidence", F),
     checkboxInput(ns("pval"), "Show p-value(log-rank test)", T),
     checkboxInput(ns("table"), "Show table", T),
+    checkboxInput(ns("number_censored"), "Show number censored", F),
     checkboxInput(ns("ci"), "Show 95% CI", F),
     checkboxInput(ns("marks"), "Show censoring marks", F),
     uiOutput(ns("ranges")),
@@ -752,7 +753,7 @@ kaplanModule <- function(input, output, session, data, data_label, data_varStruc
         if (input$cmp_risk_check) {
           return(
             jskm::jskm(res.km,
-              pval = input$pval, marks = input$marks, table = input$table, ylab = ylab, ystrataname = yst.name, ystratalabs = yst.lab, ci = input$ci, timeby = input$timeby, xlims = input$xlims, ylims = input$ylims,
+              pval = input$pval, marks = input$marks, table = input$table, table.censor = input$number_censored, ylab = ylab, ystrataname = yst.name, ystratalabs = yst.lab, ci = input$ci, timeby = input$timeby, xlims = input$xlims, ylims = input$ylims,
               cumhaz = input$cumhaz, cluster.option = "None", cluster.var = NULL, data = data.km, pval.coord = pval.coord, legendposition = legend.p, linecols = pal, xlabs = text.x, dashed = dashed, cut.landmark = cut.landmark,
               showpercent = input$showpercent, surv.scale = surv.scale, status.cmprsk = status_cmprsk, linewidth = input$linewidth_km
             )
@@ -760,7 +761,7 @@ kaplanModule <- function(input, output, session, data, data_label, data_varStruc
         } else {
           return(
             jskm::jskm(res.km,
-              pval = input$pval, marks = input$marks, table = input$table, ylab = ylab, ystrataname = yst.name, ystratalabs = yst.lab, ci = input$ci, timeby = input$timeby, xlims = input$xlims, ylims = input$ylims,
+              pval = input$pval, marks = input$marks, table = input$table, table.censor = input$number_censored, ylab = ylab, ystrataname = yst.name, ystratalabs = yst.lab, ci = input$ci, timeby = input$timeby, xlims = input$xlims, ylims = input$ylims,
               cumhaz = input$cumhaz, cluster.option = "None", cluster.var = NULL, data = data.km, pval.coord = pval.coord, legendposition = legend.p, linecols = pal, xlabs = text.x, dashed = dashed, cut.landmark = cut.landmark,
               showpercent = input$showpercent, surv.scale = surv.scale, status.cmprsk = status_cmprsk, linewidth = input$linewidth_km
             )
@@ -769,7 +770,7 @@ kaplanModule <- function(input, output, session, data, data_label, data_varStruc
       } else {
         return(
           jskm::jskm(res.km,
-            pval = input$pval, marks = input$marks, table = input$table, ylab = ylab, ystrataname = yst.name, ystratalabs = yst.lab, ci = input$ci, timeby = input$timeby, xlims = input$xlims, ylims = input$ylims,
+            pval = input$pval, marks = input$marks, table = input$table, table.censor = input$number_censored, ylab = ylab, ystrataname = yst.name, ystratalabs = yst.lab, ci = input$ci, timeby = input$timeby, xlims = input$xlims, ylims = input$ylims,
             cumhaz = input$cumhaz, cluster.option = "cluster", cluster.var = id.cluster(), data = data.km, pval.coord = pval.coord, legendposition = legend.p, linecols = pal, xlabs = text.x, dashed = dashed, cut.landmark = cut.landmark,
             showpercent = input$showpercent, surv.scale = surv.scale, status.cmprsk = status_cmprsk, linewidth = input$linewidth_km
           )
@@ -778,7 +779,7 @@ kaplanModule <- function(input, output, session, data, data_label, data_varStruc
     } else {
       return(
         jskm::svyjskm(res.km,
-          pval = input$pval, table = input$table, ylab = ylab, ystrataname = yst.name, ystratalabs = yst.lab, ci = input$ci, timeby = input$timeby, xlims = input$xlims, ylims = input$ylims,
+          pval = input$pval, table = input$table, table.censor = input$number_censored, ylab = ylab, ystrataname = yst.name, ystratalabs = yst.lab, ci = input$ci, timeby = input$timeby, xlims = input$xlims, ylims = input$ylims,
           cumhaz = input$cumhaz, design = data.km, pval.coord = pval.coord, legendposition = legend.p, linecols = pal, xlabs = text.x, dashed = dashed, cut.landmark = cut.landmark,
           showpercent = input$showpercent, surv.scale = surv.scale, linewidth = input$linewidth_km
         )
@@ -862,7 +863,8 @@ kaplanModule <- function(input, output, session, data, data_label, data_varStruc
           min = 0, max = 1, value = 0.8
         ),
         textInput(session$ns("xaxis_km"), "x-axis text", "Time-to-event"),
-        radioButtons(session$ns("pal_km"), "Line color", choices = rownames(RColorBrewer::brewer.pal.info), selected = "Set1", inline = T),
+        # radioButtons(session$ns("pal_km"), "Line color", choices = rownames(RColorBrewer::brewer.pal.info), selected = "Set1", inline = T),
+        radioButtons(session$ns("pal_km"), "Line color", choices = c("Set1", "black", "npg", "aaas", "nejm", "lancet", "jama", "jco", "frontiers"), selected = "Set1", inline = T),
         sliderInput(session$ns("linewidth_km"), "Line Width", min = 0, max = 2, value = 1, step = 0.1)
       )
     } else {
@@ -882,7 +884,8 @@ kaplanModule <- function(input, output, session, data, data_label, data_varStruc
           min = 0, max = 1, value = 0.1 + input$ylims[1]
         ),
         h3("Line"),
-        radioButtons(session$ns("pal_km"), "Line color", choices = c("black", rownames(RColorBrewer::brewer.pal.info)), selected = "Set1", inline = T),
+        # radioButtons(session$ns("pal_km"), "Line color", choices = c("black", rownames(RColorBrewer::brewer.pal.info)), selected = "Set1", inline = T),
+        radioButtons(session$ns("pal_km"), "Line color", choices = c("Set1", "black", "npg", "aaas", "nejm", "lancet", "jama", "jco", "frontiers"), selected = "Set1", inline = T),
         checkboxInput(session$ns("linetype"), "Different line type", value = F),
         h3("Others"),
         textInput(session$ns("xaxis_km"), "x-axis text", "Time-to-event"),